🆕 [#56] reinstate User Accounts Switchboard plug (ugly hack)
Due to libgnome-desktop-3 being missing from a repo, we couldn't use packages to get this plug working. However, I pulled these files right from f36 and everything seems to work absolutely fine. This is almost certainly a hacky solution and we should aim to remove it as soon as possible if we do push it into production.